feat: :sparkles: Add functionality for adding products to the shopping cart
Implemented the feature to allow users to click an "add to cart" button for a chosen product, dynamically adding it to the shopping cart. Introduced an initial state with an empty cart icon and a message indicating no products have been added yet. Enabled users to remove products from the cart by clicking the "x" icon next to each product. Added alerts to notify users when a product is added or removed from the cart. Updated the "add to cart" button icon to reflect when a product has already been added. And created a new array to store products added to the cart, which helps manage the empty cart state dynamically as products are added and removed. This is handled with conditional logic.