feat: :sparkles: Add event listeners to the "add to cart" button on the product details aside
A variable was created to store the current event listener that has the add product to shopping cart button in the product details handle of a specific product. An id was added to each product and an object was created to store the products that were added to the shopping cart by their id and it was established whether they were true or false depending on whether they were in the shopping cart or not. All of the above was done to be able to handle the events of the add product to shopping cart button from the product details side. Finally, the code was created to activate or deactivate the button when necessary and a function to add each product to the shopping cart